+++ src/main/java/davethebrave/Launcher.java
@@ -4,6 +4,6 @@ import javafx.application.Application;
 
 public class Launcher {
     public static void main(String[] args) {
-        Application.launch(DaveTheBrave.class, args);
+        Application.launch(Main.class, args);
     }
 }
