+++ src/main/java/davethebrave/ui/DialogBox.java
@@ -1,6 +1,11 @@
 package davethebrave.ui;
 
+import javafx.beans.Observable;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
 import javafx.geometry.Pos;
+import javafx.scene.Node;
+import javafx.scene.control.Dialog;
 import javafx.scene.control.Label;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
@@ -10,6 +15,23 @@ public class DialogBox extends HBox {
 
     private Label text;
     private ImageView displayPicture;
+
+    private void flip() {
+        this.setAlignment(Pos.TOP_LEFT);
+        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
+        FXCollections.reverse(tmp);
+        this.getChildren().setAll(tmp);
+    }
+
+    public static DialogBox getUserDialog(String s, Image i) {
+        return new DialogBox(s, i);
+    }
+
+    public static DialogBox getDaveDialog(String s, Image i) {
+        var db = new DialogBox(s, i);
+        db.flip();
+        return db;
+    }
 
     public DialogBox(String s, Image i) {
         text = new Label(s);
